@@ -5,7 +5,7 @@ await esbuild.build({
   bundle: true,
   format: "esm",
   target: "es2020",
-  outdir: "dist",
+  outdir: "../DualDrill.Server/wwwroot/js/dist/",
   loader: {
     ".wgsl": 'text'
   }
