@@ -1,8 +1,8 @@
 ﻿using DualDrill.Engine;
+using DualDrill.Engine.Media;
 using DualDrill.Server.Services;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.SignalR;
-using SIPSorcery.net.RTP;
 using SIPSorcery.Net;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
@@ -19,10 +19,11 @@ public interface IDrillHubClient
 }
 
 sealed class DrillHub(
-    RTCDemoVideoSource VideoSource,
+    HeadlessSurfaceCaptureVideoSource VideoSource,
     FrameSimulationService UpdateService,
     ILogger<DrillHub> Logger,
-    IHubContext<DrillHub, IDrillHubClient> HubContext) : Hub<IDrillHubClient>
+    IHubContext<DrillHub, IDrillHubClient> HubContext,
+    RTCPeerConnectionProviderService RTCPeerConnectionProviderService) : Hub<IDrillHubClient>
 {
     static readonly string RTCConnectionKey = "RTCConnection";
 
@@ -30,6 +31,11 @@ sealed class DrillHub(
     {
         await base.OnConnectedAsync();
         Logger.LogInformation("{ConnectionId} Connected", Context.ConnectionId);
+        if (RTCPeerConnection is null)
+        {
+            var pc = RTCPeerConnectionProviderService.CreatePeerConnection(Context.ConnectionId);
+            RTCPeerConnection = pc;
+        }
     }
 
     public async Task AddIceCandidate(string data)
@@ -41,81 +47,35 @@ sealed class DrillHub(
         }
     }
 
-    public async ValueTask<string> CreatePeerConnection(string offer)
+    public async ValueTask<string> Negotiate(string offer)
     {
-        var connectionId = Context.ConnectionId;
-        var pc = new RTCPeerConnection(new RTCConfiguration
-        {
-            iceServers = new List<RTCIceServer>()
-        });
-        Context.Items.Add(RTCConnectionKey, pc);
-        var track = new MediaStreamTrack(VideoSource.VideoEncoder.SupportedFormats, MediaStreamStatusEnum.SendOnly);
-        pc.addTrack(track);
-        pc.setRemoteDescription(new RTCSessionDescriptionInit
-        {
-            type = RTCSdpType.offer,
-            sdp = offer,
-        });
-        pc.oniceconnectionstatechange += (state) =>
-         {
-             Logger.LogWarning($"ice candidate state changed to {Enum.GetName(state)}");
-         };
-        pc.onicecandidateerror += (err, msg) =>
-        {
-            Logger.LogError(msg + Environment.NewLine + JsonSerializer.Serialize(err));
-        };
-        pc.onnegotiationneeded += () =>
-        {
-            Logger.LogWarning("Negotiation Needed");
-        };
-        var answer = pc.createAnswer(new RTCAnswerOptions { X_ExcludeIceCandidates = true });
-        await pc.setLocalDescription(answer);
-
-        pc.onicecandidate += (c) =>
-        {
-            HubContext.Clients.Client(connectionId).IceCandidate(JsonSerializer.Serialize(c));
-        };
-
-        void sendVideo(uint duration, VideoFrameBuffer buffer)
-        {
-            Task.Run(() =>
-            {
-                if (pc.connectionState == RTCPeerConnectionState.connected)
-                {
-                    var s = pc.VideoStreamList.First(s => s.LocalTrack == track);
-                    s.SendVideo(duration, buffer.Memory.ToArray());
-                }
-            });
-        }
-
-        pc.onconnectionstatechange += async (state) =>
-        {
-            Console.WriteLine($"Peer connection state change to {state}.");
-
-            switch (state)
-            {
-                case RTCPeerConnectionState.connected:
-                    VideoSource.OnVideoSourceEncodedSample += sendVideo;
-                    Logger.LogInformation($"{connectionId} RTC Connected");
-                    break;
-                case RTCPeerConnectionState.failed:
-                    pc.Close("ice disconnection");
-                    break;
-                case RTCPeerConnectionState.closed:
-                    VideoSource.OnVideoSourceEncodedSample -= sendVideo;
-                    Logger.LogInformation($"{connectionId} RTC Disconnected");
-                    break;
-            }
-        };
-
-        return answer.sdp;
+        Logger.LogInformation("Negotate called for {connectionId}", Context.ConnectionId);
+        return await RTCPeerConnectionProviderService.negotiateAsync(Context.ConnectionId, RTCPeerConnection, offer);
     }
 
     public RTCPeerConnection? RTCPeerConnection
     {
         get
         {
-            return Context.Items[RTCConnectionKey] as RTCPeerConnection;
+            if (Context.Items.TryGetValue(RTCConnectionKey, out var pc))
+            {
+                return pc as RTCPeerConnection;
+            }
+            else
+            {
+                return null;
+            }
+        }
+        set
+        {
+            if (Context.Items.TryGetValue(RTCConnectionKey, out _))
+            {
+                Context.Items[RTCConnectionKey] = value;
+            }
+            else
+            {
+                Context.Items.Add(RTCConnectionKey, value);
+            }
         }
     }
 
