@@ -1,16 +1,17 @@
 ﻿using DualDrill.Engine;
-using DualDrill.Graphics.Headless;
+using DualDrill.Engine.Headless;
+using DualDrill.Engine.Media;
 using Microsoft.Extensions.Options;
 using System.Threading.Channels;
 
 namespace DualDrill.Server;
 
-public sealed class HeadlessRealtimeFrameHostedService : BackgroundService
+public sealed class RealtimeFrameHostedService : BackgroundService
 {
     readonly TimeSpan SampleRate = TimeSpan.FromSeconds(1.0 / 60.0);
     readonly TimeProvider TimeProvider = TimeProvider.System;
 
-    private readonly ILogger<HeadlessRealtimeFrameHostedService> Logger;
+    private readonly ILogger<RealtimeFrameHostedService> Logger;
     private readonly IFrameService FrameService;
     private readonly Channel<int> FrameChannel;
 
@@ -18,20 +19,24 @@ public sealed class HeadlessRealtimeFrameHostedService : BackgroundService
 
     private readonly HeadlessSurface Surface;
 
-    public HeadlessRealtimeFrameHostedService(
+    public HeadlessSurfaceCaptureVideoSource VideoSource { get; }
+
+    public RealtimeFrameHostedService(
         IFrameService frameService,
-        ILogger<HeadlessRealtimeFrameHostedService> logger,
-        HeadlessSurface surface)
+        ILogger<RealtimeFrameHostedService> logger,
+        HeadlessSurface surface,
+        HeadlessSurfaceCaptureVideoSource videoSource)
     {
         Logger = logger;
         FrameService = frameService;
         FrameChannel = Channel.CreateBounded<int>(1);
         Surface = surface;
+        VideoSource = videoSource;
     }
 
     private static void TimerFrameCallback(object? data)
     {
-        var self = (HeadlessRealtimeFrameHostedService)data!;
+        var self = (RealtimeFrameHostedService)data!;
         if (!self.FrameChannel.Writer.TryWrite(self.FrameIndex))
         {
             self.Logger.LogWarning("Frame skipped {CurrentFrame}", self.FrameIndex);
@@ -41,17 +46,23 @@ public sealed class HeadlessRealtimeFrameHostedService : BackgroundService
 
     protected override async Task ExecuteAsync(CancellationToken stoppingToken)
     {
+        await Task.Yield();
+        await VideoSource.StartVideo();
         using var timer = TimeProvider.CreateTimer(TimerFrameCallback, this, TimeSpan.Zero, SampleRate);
         await foreach (var frameIndex in FrameChannel.Reader.ReadAllAsync(stoppingToken))
         {
-            _ = Surface.TryAcquireImage(frameIndex);
+            var surfaceImage = Surface.TryAcquireImage();
+            if (surfaceImage is null)
+            {
+                Logger.LogWarning("Failed to get render target from headless surface");
+            }
             await FrameService.OnFrameAsync(new FrameContext
             {
                 FrameIndex = frameIndex,
                 MouseEvent = (MouseEvent[])[],
                 Surface = Surface,
             }, stoppingToken);
-            await Surface.PresentAsync(stoppingToken);
+            Surface.Present();
         }
     }
 }
